ещё немного активитей

+++ app/src/main/java/com/example/Classes/API/JsonGetBarcodeByCouponApi.java
@@ -8,6 +8,6 @@ import retrofit2.http.Path;
 
 public interface JsonGetBarcodeByCouponApi {
     @GET("/api/barcode/get/by_coupon/{id}")
-    Call<BarCode> getBarcodeWithId(@Path("id")int id);
+    Call<BarCode> getBarcodeWithId(@Path("id") Long id);
 }
